秒杀修改

+++ jmp/mp-service/src/main/java/com/vpu/mp/service/pojo/shop/market/seckill/SeckillProductAddParam.java
@@ -16,5 +16,5 @@ public class SeckillProductAddParam {
     @NotNull
     private BigDecimal secKillPrice;
     @NotNull
-    private Short      stock;
+    private Integer      stock;
 }
